trips folder is better organized on a component basis
