+++ src/app/trips/index.ts
@@ -3,9 +3,9 @@ export * from './create-trip/index'
 export * from './trip-overview/index'
 
 // components
-export * from './trips-list.component'
-export * from './trips-title-section.component'
-export * from './trip-thumbnail.component'
-export * from './new-trip-thumbnail.component'
-export * from './trips-list-resolver.component'
+export * from './trips-list/trips-list.component'
+export * from './trips-title-section/trips-title-section.component'
+export * from './trip-thumbnail/trip-thumbnail.component'
+export * from './new-trip-thumbnail/new-trip-thumbnail.component'
+export * from './trips-list/trips-list-resolver.component'
