+++ src/app/trips/new-trip-thumbnail/new-trip-thumbnail.component.css
