+++ src/app/trips/trip-thumbnail/trip-thumbnail.component.html
