+++ src/app/trips/trips-list/trips-list.component.css
