+++ src/styles.css
@@ -2,10 +2,6 @@
 @import "~bootstrap-icons/font/bootstrap-icons.css";
 @import url('https://fonts.googleapis.com/css2?family=Sura:wght@400;700&display=swap');
 
-/* prevents any */
-body {
-}
-
 body {
   color: rgb(31, 31, 31);
   background-color: rgb(245, 245, 245);
